fix(event-listener): move non-services folders to src/ root
