+++ packages/event-listener/src/email/IEmailAdapter.ts
@@ -1,4 +1,4 @@
-import { IEmail, IEmailResponse } from '../email.service';
+import { IEmail, IEmailResponse } from '../services/email.service';
 
 export interface IEmailAdapter {
     send(from: string, email: IEmail): Promise<IEmailResponse>;
