+++ packages/event-listener/src/services/email.service.ts
@@ -1,4 +1,4 @@
-import { IEmailAdapter } from './email/IEmailAdapter';
+import { IEmailAdapter } from '../email/IEmailAdapter';
 
 export interface IEmail {
     to: string[];
