+++ packages/event-listener/test/helpers/TestAdapter.ts
@@ -1,11 +1,8 @@
 import { IEmail, IEmailResponse } from '../../src/services/email.service';
-import { IEmailAdapter } from '../../src/services/email/IEmailAdapter';
+import { IEmailAdapter } from '../../src/email/IEmailAdapter';
 
 export class TestEmailAdapter implements IEmailAdapter {
-    public async send(
-        from: string,
-        email: IEmail
-    ): Promise<IEmailResponse> {
+    public async send(from: string, email: IEmail): Promise<IEmailResponse> {
         console.log(`Sent test email from ${from}`);
         console.log(email);
